protection against wrong oversampling value

+++ lib/segment/src/types.rs
@@ -277,7 +277,7 @@ pub struct QuantizationSearchParams {
     /// For example, if `oversampling` is 2.4 and `limit` is 100, then 240 vectors will be pre-selected using quantized index,
     /// and then top-100 will be returned after re-scoring.
     #[serde(default = "default_quantization_oversampling_value")]
-    pub oversampling: Option<f32>,
+    pub oversampling: Option<f64>,
 }
 
 pub const fn default_quantization_ignore_value() -> bool {
@@ -288,7 +288,7 @@ pub const fn default_quantization_rescore_value() -> bool {
     false
 }
 
-pub const fn default_quantization_oversampling_value() -> Option<f32> {
+pub const fn default_quantization_oversampling_value() -> Option<f64> {
     None
 }
